spring-boot-apiVersion-版本float精确匹配

+++ spring-boot-demos/src/main/java/com/mk/demos/spring/boot/annotation/ApiVersion.java
@@ -23,5 +23,5 @@ public @interface ApiVersion {
     /**
      * @return 版本号
      */
-    int value() default 1;
+    float value() default 1f;
 }
